Restructure app into modular backend and Vite frontend
