+++ .gitignore
@@ -1,0 +1,4 @@
+node_modules/
+ui/node_modules/
+ui/dist/
+npm-debug.log*
