+++ README.md
@@ -1,51 +1,106 @@
-# terminal-worktree CLI
+# terminal-worktree
 
-This repository hosts the early prototype of the `terminal-worktree` console application. The first milestone is a small CLI that serves the static UI mockup in `ui.sample.html` so that it is accessible in a browser.
+`terminal-worktree` is a password-protected CLI that serves a browser-based interface for managing
+Git repositories, worktrees, and persistent shell sessions. The backend is an ES module Node.js
+service that exposes REST + WebSocket APIs, while the frontend is a Vite-powered React application
+that lives in `ui/`.
 
 ## Prerequisites
 
-- Node.js 18 or newer (the CLI uses the built-in HTTP server and modern JavaScript features).
+- Node.js 18 or newer
+- Git (for cloning repositories and creating worktrees)
 
-## Usage
+## Install Dependencies
 
-From the repository root run:
+From the repository root:
+
+```bash
+npm install
+```
+
+Install frontend dependencies separately:
+
+```bash
+cd ui
+npm install
+```
+
+## Build the Frontend
+
+The backend serves pre-built assets from `ui/dist`. Build them once (or whenever the UI changes):
+
+```bash
+npm run build          # Invokes `vite build` in ui/
+```
+
+The build artefacts land in `ui/dist/` and are picked up automatically by the CLI defaults.
+
+## Running the CLI
+
+After building the UI:
+
+```bash
+npm start
+```
+
+or directly:
 
 ```bash
 node bin/terminal-worktree.js
 ```
 
-By default the UI is served from `ui.sample.html` on port `3414` and bound to `0.0.0.0`. Customise this behaviour with flags:
+By default the server binds to `0.0.0.0:3414`, serves assets from `ui/dist`, and scans the current
+working directory for repositories. Customise behaviour with flags:
 
 ```bash
-node bin/terminal-worktree.js --port 4000 --host 127.0.0.1 --ui path/to/file.html
+node bin/terminal-worktree.js \
+  --port 4000 \
+  --host 127.0.0.1 \
+  --ui ./ui/dist \
+  --workdir /path/to/workdir \
+  --password secret
 ```
 
 ### CLI Options
 
-- `-p, --port <number>`: Port for the HTTP server (default: `3414`).
-- `-H, --host <host>`: Host/interface to bind (default: `0.0.0.0`).
-- `-u, --ui <path>`: Path to the UI HTML file to serve (default: `ui.sample.html`).
-- `-w, --workdir <path>`: Root directory that contains cloned repositories (default: current working directory).
-- `-P, --password <string>`: Password that guards the web UI (default: randomly generated at startup).
-- `-h, --help`: Display the usage information.
-- `-v, --version`: Print the package version.
+- `-p, --port <number>` – HTTP port (default: `3414`)
+- `-H, --host <host>` – Bind address (default: `0.0.0.0`)
+- `-u, --ui <path>` – Directory or entry file for the built UI (default: `ui/dist`)
+- `-w, --workdir <path>` – Root directory that holds `org/repo` folders (default: process CWD)
+- `-P, --password <string>` – UI password (default: secure random string printed at startup)
+- `-h, --help` – Print usage
+- `-v, --version` – Show package version
 
 ### Authentication
 
-Every time the server boots it prints the password that protects the web UI. If you do not provide
-one with `--password`, a secure 12-character password is generated and written to stdout alongside
-the usual startup logs. The browser prompts for this password before any repository APIs are
-available. Successful logins receive an HTTP-only session cookie; click **Log out** in the UI (or hit
-`POST /api/auth/logout`) to terminate the session and force the login screen to reappear.
+Every server boot prints the UI password. Clients must authenticate before calling API endpoints.
+Successful logins receive an HTTP-only session cookie; log out via the UI or `POST
+/api/auth/logout`. On shutdown the backend cleans up shell sessions, tmux attachments, and WebSocket
+clients.
 
-The server keeps running until you terminate the process (e.g. `Ctrl+C`). The mock UI loads React
-and supporting libraries from public CDNs, so make sure the machine running the browser has
-internet access. Repository entries in the sidebar are populated from the given work directory
-following the structure `[workdir]/org-name/repo-name/repository`. Using the **Add Repo** button
-will clone the given repository URL into that structure automatically, and the embedded terminal
-connects over WebSockets to a `node-pty` session so aliases, login scripts, and interactive tools
-behave exactly as they do in your local shell. The default `main` branch is visible but intentionally
-blocked from opening a terminal session or being deleted. Each new worktree is created only after the
-`repository` checkout pulls the latest `origin/main`, ensuring fresh state before branching. When you
-open a worktree for the first time the UI offers quick actions (Terminal, Codex, Cursor, Claude) that
-automatically bootstrap the appropriate tool; subsequent visits reconnect to the existing shell.
+### Repository Layout & Worktrees
+
+The work directory is expected to follow:
+
+```
+[workdir]/
+  org/
+    repo/
+      repository/      # main checkout
+      <worktree-name>/ # additional worktrees
+```
+
+Use **Add Repo** to clone into this structure, **Create Worktree** to branch from up-to-date
+`origin/main`, and **Delete Worktree** to remove worktrees (except `main`, which is intentionally
+protected). Terminal sessions are backed by `node-pty` and optionally tmux so reconnects resume the
+previous shell.
+
+## Development Workflow
+
+- `npm run dev` – Start the backend CLI.
+- `npm run dev:ui` – Run the Vite dev server (hot-module reloading React).
+- `npm run preview` – Serve the production build through Vite.
+- `npm run build` – Generate production assets in `ui/dist`.
+
+During local development run the backend CLI (`npm run dev`) alongside the Vite dev server (`npm run
+dev:ui`). Use the dev server URL directly in the browser while exercising APIs exposed by the CLI.
