+++ bin/terminal-worktree.js
@@ -1,5 +1,5 @@
 #!/usr/bin/env node
 
-const { main } = require('../src/cli.js');
+import { main } from '../src/cli.js';
 
 main();
