+++ src/utils/http.js
@@ -1,0 +1,43 @@
+import { MAX_REQUEST_BODY_SIZE } from '../config/constants.js';
+
+export function sendJson(res, statusCode, payload) {
+  res.statusCode = statusCode;
+  res.setHeader('Content-Type', 'application/json; charset=utf-8');
+  res.setHeader('Cache-Control', 'no-store');
+  res.end(JSON.stringify(payload));
+}
+
+export async function readJsonBody(req) {
+  return new Promise((resolve, reject) => {
+    const chunks = [];
+    let length = 0;
+
+    req.on('data', (chunk) => {
+      chunks.push(chunk);
+      length += chunk.length;
+
+      if (length > MAX_REQUEST_BODY_SIZE) {
+        reject(new Error('Request body too large'));
+        req.destroy();
+      }
+    });
+
+    req.on('end', () => {
+      if (chunks.length === 0) {
+        resolve({});
+        return;
+      }
+
+      try {
+        const buffer = Buffer.concat(chunks);
+        resolve(JSON.parse(buffer.toString('utf8')));
+      } catch (error) {
+        reject(new Error('Invalid JSON payload'));
+      }
+    });
+
+    req.on('error', (error) => {
+      reject(error);
+    });
+  });
+}
