+++ ui/src/styles.css
@@ -1,0 +1,19 @@
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
+
+@layer base {
+  html,
+  body,
+  #root {
+    @apply h-full;
+  }
+
+  body {
+    @apply m-0 bg-neutral-950 text-neutral-100;
+  }
+}
+
+.xterm .xterm-viewport {
+  @apply bg-neutral-950;
+}
