+++ ui/tailwind.config.cjs
@@ -1,0 +1,14 @@
+module.exports = {
+  content: ['./index.html', './src/**/*.{js,jsx,ts,tsx}'],
+  darkMode: 'class',
+  theme: {
+    extend: {
+      colors: {
+        neutral: {
+          925: '#161616',
+        },
+      },
+    },
+  },
+  plugins: [],
+};
